LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/values/shape-outside-circle-000-expected.txt


FAIL No arguments - inline assert_equals: expected "circle()" but got "circle(at 50% 50%)"
FAIL Radial argument only - inline assert_equals: expected "circle(50px)" but got "circle(50px at 50% 50%)"
PASS Position argument only - inline
PASS Radial and position argument - inline
FAIL No arguments - computed assert_equals: expected "circle()" but got "circle(at 50% 50%)"
FAIL Radial argument only - computed assert_equals: expected "circle(50px)" but got "circle(50px at 50% 50%)"
PASS Position argument only - computed
PASS Radial and position argument - computed
